Created a semi-working toggle button
