+++ app/src/main/java/org/schabi/newpipe/local/subscription/item/FeedGroupCardVerticalItem.kt
@@ -1,7 +1,12 @@
 package org.schabi.newpipe.local.subscription.item
 
+import android.view.MotionEvent
 import android.view.View
+import androidx.recyclerview.widget.ItemTouchHelper
+import androidx.recyclerview.widget.ItemTouchHelper.DOWN
+import androidx.recyclerview.widget.ItemTouchHelper.UP
 import com.xwray.groupie.viewbinding.BindableItem
+import com.xwray.groupie.viewbinding.GroupieViewHolder
 import org.schabi.newpipe.R
 import org.schabi.newpipe.database.feed.model.FeedGroupEntity
 import org.schabi.newpipe.databinding.FeedGroupCardVerticalItemBinding
@@ -10,9 +15,11 @@ import org.schabi.newpipe.local.subscription.FeedGroupIcon
 data class FeedGroupCardVerticalItem(
     val groupId: Long = FeedGroupEntity.GROUP_ALL_ID,
     val name: String,
-    val icon: FeedGroupIcon
+    val icon: FeedGroupIcon,
+    val dragCallback: ItemTouchHelper
 ) : BindableItem<FeedGroupCardVerticalItemBinding>() {
-    constructor (feedGroupEntity: FeedGroupEntity) : this(feedGroupEntity.uid, feedGroupEntity.name, feedGroupEntity.icon)
+    constructor (feedGroupEntity: FeedGroupEntity, dragCallback: ItemTouchHelper) :
+        this(feedGroupEntity.uid, feedGroupEntity.name, feedGroupEntity.icon, dragCallback)
 
     override fun getId(): Long {
         return when (groupId) {
@@ -26,6 +33,22 @@ data class FeedGroupCardVerticalItem(
     override fun bind(viewBinding: FeedGroupCardVerticalItemBinding, position: Int) {
         viewBinding.title.text = name
         viewBinding.icon.setImageResource(icon.getDrawableRes())
+    }
+
+    override fun bind(viewHolder: GroupieViewHolder<FeedGroupCardVerticalItemBinding>, position: Int, payloads: MutableList<Any>) {
+        super.bind(viewHolder, position, payloads)
+        viewHolder.binding.handle.setOnTouchListener { _, event ->
+            if (event.actionMasked == MotionEvent.ACTION_DOWN) {
+                dragCallback.startDrag(viewHolder)
+                return@setOnTouchListener true
+            }
+
+            false
+        }
+    }
+
+    override fun getDragDirs(): Int {
+        return UP or DOWN
     }
 
     override fun initializeViewBinding(view: View) = FeedGroupCardVerticalItemBinding.bind(view)
