--- app/src/main/res/layout/feed_group_reorder_item.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginLeft="8dp"
-    android:layout_marginTop="3dp"
-    android:layout_marginRight="8dp"
-    android:layout_marginBottom="3dp"
-    android:minHeight="?listPreferredItemHeightSmall"
-    android:orientation="horizontal"
-    app:cardBackgroundColor="?attr/card_item_background_color"
-    app:cardCornerRadius="5dp"
-    app:cardElevation="2dp">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="center_vertical">
-
-        <androidx.appcompat.widget.AppCompatImageView
-            android:id="@+id/group_icon"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:layout_alignParentLeft="true"
-            android:layout_centerVertical="true"
-            android:layout_marginLeft="16dp"
-            tools:ignore="ContentDescription,RtlHardcoded"
-            tools:src="@drawable/ic_kiosk_hot" />
-
-        <org.schabi.newpipe.views.NewPipeTextView
-            android:id="@+id/group_name"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_centerVertical="true"
-            android:layout_marginLeft="16dp"
-            android:layout_marginTop="6dp"
-            android:layout_marginBottom="6dp"
-            android:layout_toLeftOf="@+id/handle"
-            android:layout_toRightOf="@+id/group_icon"
-            android:ellipsize="end"
-            android:maxLines="2"
-            android:textAppearance="?textAppearanceListItem"
-            tools:ignore="RtlHardcoded"
-            tools:text="Lorem ipsum dolor sit amet" />
-
-        <androidx.appcompat.widget.AppCompatImageView
-            android:id="@+id/handle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentRight="true"
-            android:layout_centerVertical="true"
-            android:paddingLeft="16dp"
-            android:paddingTop="12dp"
-            android:paddingRight="16dp"
-            android:paddingBottom="12dp"
-            android:src="@drawable/ic_drag_handle"
-            tools:ignore="ContentDescription,RtlHardcoded" />
-    </RelativeLayout>
-</androidx.cardview.widget.CardView>
